fix: Debug and fix VAPI webhook integration for real analytics
- Fix VAPI webhook payload parsing: extract data from 'message' wrapper for server-url webhooks
- Add parse_timestamp support for integer millisecond timestamps and ISO strings
- Change timestamp_ms column from Integer to BigInteger to handle large timestamps
- Add handle_conversation_update to store transcripts from conversation-update events
- Add fallback analysis trigger in handle_status_update when call ends
- Fix timezone-aware datetime handling (use datetime.now(timezone.utc) instead of utcnow)
- Fix API response to exclude non-existent model attributes (previous_score, score_change)
- Add signal attributes to API response (dsm5_criteria, verbatim_quote, quote_context)
- Fix frontend polling cleanup: add isPollingRef and isMountedRef to prevent stale calls
- Reduce backend logging noise: only log important webhook events, silence noisy ones
- Disable SQL echo logging in database.py to reduce verbose output

These changes enable the complete voice session analytics pipeline:
Call ends → Transcripts stored → LLM analysis → Frontend receives real data instead of mocks

+++ backend/src/models/session.py
@@ -75,7 +75,7 @@ class Transcript(Base, TimestampMixin):
     # Content
     role: Mapped[str] = mapped_column(String(20), nullable=False)  # assistant, user
     content: Mapped[str] = mapped_column(Text, nullable=False)
-    timestamp_ms: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
+    timestamp_ms: Mapped[Optional[int]] = mapped_column(BigInteger, nullable=True)
 
     # Speech features (populated during analysis)
     speech_speed: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
